src/requirements-base.txt
keyring
azure-keyvault-secrets
azure-identity
pyyaml
pydantic
typer
tabulate
